run tests individually
